SZ: Github actions tests pipeline
