+++ package.json
@@ -575,7 +575,8 @@
     "esbuild": "npm run esbuild-base -- --sourcemap && npm run esbuild-server -- --sourcemap",
     "package": "vsce package",
     "publish": "vsce publish",
-    "test": "vitest run --reporter verbose --coverage"
+    "test": "vitest run --reporter verbose --coverage",
+    "gh-test": "vitest run --reporter=junit --outputFile=./test-output.xml --coverage"
   },
   "devDependencies": {
     "@types/glob": "^7.1.4",
