+++ tests/internal/e2e/package.json
@@ -1,29 +1,31 @@
 {
     "scripts": {
-        "test": "npx wdio run test/wdio.conf.ts"
+        "test": "npx wdio run test/wdio.conf.ts",
+        "gh-test": "Xvfb :99 -screen 0 1920x1080x24 & >> /dev/null && xvfb-run --server-num=99 --auto-servernum --server-args=\"-screen 0 1920x1080x24+32\" npx wdio run test/wdio.conf.ts"
     },
     "devDependencies": {
         "@types/global-agent": "^2.1.1",
         "@typescript-eslint/eslint-plugin": "^5.54.1",
         "@typescript-eslint/parser": "^5.54.1",
+        "@wdio/cli": "^8.6.5",
+        "@wdio/junit-reporter": "^8.32.2",
         "@wdio/local-runner": "^8.6.3",
         "@wdio/mocha-framework": "^8.6.1",
         "@wdio/spec-reporter": "^8.4.0",
         "child_process": "^1.0.2",
-        "chromedriver": "^119.0.1",
+        "chromedriver": "^107.0.3",
         "create-wdio": "^8.2.2",
         "csv-parse": "^5.3.4",
+        "eslint": "^8.35.0",
+        "eslint-config-prettier": "^8.7.0",
+        "eslint-plugin-prettier": "^4.2.1",
         "global-agent": "^2.2.0",
         "path": "^0.12.7",
         "ts-node": "^10.9.1",
         "typescript": "^4.9.5",
+        "vscode-uri": "3.0.7",
         "wdio-chromedriver-service": "^8.1.1",
         "wdio-devtools-service": "^0.1.6",
-        "wdio-vscode-service": "^5.0.0",
-        "@wdio/cli": "^8.6.5",
-        "eslint": "^8.35.0",
-        "eslint-config-prettier": "^8.7.0",
-        "eslint-plugin-prettier": "^4.2.1",
-        "vscode-uri":"3.0.7"
+        "wdio-vscode-service": "^5.0.0"
     }
 }
